Make the gem usable in stand-alone scripts
Why?
----

When trying to use the gem from within a Ruby script we get an
error indicating that the constant `JSON` is not initialized.

**Note**: This error doesn't happen if the gem is used within a
Rails app because of Rails' auto-loading feature that already
load `json` into memory.

What?
-----

Add the missing `require` statement.

How?
----

Add a `require` statement for `'json'` to
`easy_broker.rb`.